fix: Correct link to code samples at bottom of page (#290)
A recent edit pointed the **Examples** link in the navbar to a new location, but it looks as if the link at the bottom of the page was overlooked. (It takes you to the boneyard.) This edit updates the href at the bottom to match the href in the navbar.